Add experience balance

+++ Assets/Source/Scripts/Loot/DiamondLootFactory.cs
@@ -5,5 +5,5 @@ public class DiamondLootFactory : LootFactory
     }
 
     protected override string AssetKey => AssetsKeys.DiamondLoot;
-    protected override int BaseExperienceCount => 3;
+    protected override float BaseExperienceCount => 3;
 }
